requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
httpx==0.25.2

# JWT Authentication
pyjwt[crypto]==2.8.0
cryptography>=41.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
structlog==23.2.0

# Environment
python-dotenv==1.0.0